+++ odd_collector/adapters/mongodb/exceptions.py
